# Add option to speedup Botan DRNG backend with ChaCha20