--- a/common/meson.build
+++ b/common/meson.build
@@ -64,13 +64,8 @@ else
 	conf_data.set('ESDM_BOTAN', false)
 endif
 
-if get_option('botan_drng_type') == 'chacha20'
-	conf_data.set('ESDM_BOTAN_DRNG_CHACHA20', true)
-	conf_data.set('ESDM_BOTAN_DRNG_HMAC', false)
-else
-	conf_data.set('ESDM_BOTAN_DRNG_CHACHA20', false)
-	conf_data.set('ESDM_BOTAN_DRNG_HMAC', true)
-endif
+conf_data.set('ESDM_BOTAN_DRNG_CHACHA20', get_option('botan_drng_type') == 'chacha20')
+conf_data.set('ESDM_BOTAN_DRNG_HMAC', get_option('botan_drng_type') == 'hmac')
 
 # This option currently is not configurable!
 conf_data.set('CONFIG_ESDM_USE_PTHREAD', true)