For small requests ESDMs performance over RPC is mostly dominated by context switch overhead and system calls. Nevertheless a performance increase is possible when optionally allowing the ChaCha20-Ctr-DRNG of Botan, when in this mode.

Before (only HMAC):
```
Request size: 1 | Elapsed: 653.82ms | Rate: 0.03 MB/s | Iterations: 30589.45 1/s
Request size: 2 | Elapsed: 342.09ms | Rate: 0.12 MB/s | Iterations: 58464.88 1/s
Request size: 4 | Elapsed: 317.37ms | Rate: 0.25 MB/s | Iterations: 63018.83 1/s
Request size: 8 | Elapsed: 307.40ms | Rate: 0.52 MB/s | Iterations: 65062.40 1/s
Request size: 16 | Elapsed: 330.37ms | Rate: 0.97 MB/s | Iterations: 60537.75 1/s
Request size: 32 | Elapsed: 318.07ms | Rate: 2.01 MB/s | Iterations: 62878.98 1/s
Request size: 64 | Elapsed: 350.37ms | Rate: 3.65 MB/s | Iterations: 57082.09 1/s
Request size: 128 | Elapsed: 378.39ms | Rate: 6.77 MB/s | Iterations: 52854.83 1/s
Request size: 256 | Elapsed: 443.01ms | Rate: 11.56 MB/s | Iterations: 45146.21 1/s
Request size: 512 | Elapsed: 461.04ms | Rate: 22.21 MB/s | Iterations: 43379.74 1/s
Request size: 1024 | Elapsed: 702.47ms | Rate: 29.15 MB/s | Iterations: 28470.78 1/s
Request size: 2048 | Elapsed: 1.13s | Rate: 36.34 MB/s | Iterations: 17744.18 1/s
```

After (ChaCha20 selected):
```
Request size: 1 | Elapsed: 514.79ms | Rate: 0.04 MB/s | Iterations: 38850.79 1/s
Request size: 2 | Elapsed: 196.20ms | Rate: 0.20 MB/s | Iterations: 101939.18 1/s
Request size: 4 | Elapsed: 188.13ms | Rate: 0.43 MB/s | Iterations: 106309.59 1/s
Request size: 8 | Elapsed: 179.07ms | Rate: 0.89 MB/s | Iterations: 111688.99 1/s
Request size: 16 | Elapsed: 184.59ms | Rate: 1.73 MB/s | Iterations: 108349.60 1/s
Request size: 32 | Elapsed: 182.02ms | Rate: 3.52 MB/s | Iterations: 109876.80 1/s
Request size: 64 | Elapsed: 201.91ms | Rate: 6.34 MB/s | Iterations: 99052.68 1/s
Request size: 128 | Elapsed: 224.12ms | Rate: 11.42 MB/s | Iterations: 89237.39 1/s
Request size: 256 | Elapsed: 291.21ms | Rate: 17.58 MB/s | Iterations: 68678.94 1/s
Request size: 512 | Elapsed: 223.15ms | Rate: 45.89 MB/s | Iterations: 89626.23 1/s
Request size: 1024 | Elapsed: 241.89ms | Rate: 84.67 MB/s | Iterations: 82681.45 1/s
Request size: 2048 | Elapsed: 234.91ms | Rate: 174.37 MB/s | Iterations: 85140.77 1/s
```