@thillux 2025-03-25
Especially for the relevant 32 Byte/256 Bit, the speedup is noticeable.